Integrate Bignum specializations in IntegerNodes
* alias % to #modulo directly in Java, like it was done for Bignum but not Fixnum.
* Use a BooleanCastNode in Integer#== like it was in Bignum#==.
* Move Fixnum#divmod to Integer and rename the primitive.
* #hash for large integers are now handled by Kernel#hash, like MRI.
* Integer#inspect is just an alias of #to_s and also accepts a base argument.
  (Fixnum#inspect wrongly did not accept a base argument and is now fixed)
