+++ src/main/java/org/truffleruby/core/kernel/KernelNodes.java
@@ -718,8 +718,13 @@ public abstract class KernelNodes {
             return getContext().getHashing(this).hash(CLASS_SALT, Boolean.valueOf(value).hashCode());
         }
 
+        @Specialization(guards = "isRubyBignum(value)")
+        public long hashBignum(DynamicObject value) {
+            return getContext().getHashing(this).hash(CLASS_SALT, Layouts.BIGNUM.getValue(value).hashCode());
+        }
+
         @TruffleBoundary
-        @Specialization
+        @Specialization(guards = "!isRubyBignum(self)")
         public int hash(DynamicObject self) {
             // TODO(CS 8 Jan 15) we shouldn't use the Java class hierarchy like this - every class should define it's
             // own @CoreMethod hash
