+++ src/main/java/org/truffleruby/core/numeric/BignumNodes.java
@@ -15,25 +15,12 @@ import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.object.DynamicObject;
 import com.oracle.truffle.api.profiles.ConditionProfile;
-import org.jcodings.specific.USASCIIEncoding;
 import org.truffleruby.Layouts;
 import org.truffleruby.builtins.CoreClass;
-import org.truffleruby.builtins.CoreMethod;
 import org.truffleruby.builtins.CoreMethodArrayArgumentsNode;
 import org.truffleruby.builtins.Primitive;
 import org.truffleruby.builtins.PrimitiveArrayArgumentsNode;
-import org.truffleruby.core.CoreLibrary;
-import org.truffleruby.core.cast.BooleanCastNode;
-import org.truffleruby.core.cast.BooleanCastNodeGen;
-import org.truffleruby.core.cast.ToIntNode;
-import org.truffleruby.core.numeric.BignumNodesFactory.DivNodeFactory;
-import org.truffleruby.core.numeric.BignumNodesFactory.MulNodeFactory;
-import org.truffleruby.core.rope.CodeRange;
-import org.truffleruby.core.string.StringNodes;
-import org.truffleruby.language.NotProvided;
 import org.truffleruby.language.WarnNode;
-import org.truffleruby.language.control.RaiseException;
-import org.truffleruby.language.dispatch.CallDispatchHeadNode;
 
 import java.math.BigInteger;
 
@@ -50,615 +37,6 @@ public abstract class BignumNodes {
                 fixnumOrBignum = insert(new FixnumOrBignumNode());
             }
             return fixnumOrBignum.fixnumOrBignum(value);
-        }
-
-    }
-
-    @CoreMethod(names = "-@")
-    public abstract static class NegNode extends BignumCoreMethodNode {
-
-        @Specialization
-        public Object neg(DynamicObject value) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(value).negate());
-        }
-
-    }
-
-    @CoreMethod(names = "+", required = 1)
-    public abstract static class AddNode extends BignumCoreMethodNode {
-
-        @Specialization
-        public Object add(DynamicObject a, long b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).add(BigInteger.valueOf(b)));
-        }
-
-        @Specialization
-        public double add(DynamicObject a, double b) {
-            return Layouts.BIGNUM.getValue(a).doubleValue() + b;
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public Object add(DynamicObject a, DynamicObject b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).add(Layouts.BIGNUM.getValue(b)));
-        }
-
-    }
-
-    @CoreMethod(names = "-", required = 1)
-    public abstract static class SubNode extends BignumCoreMethodNode {
-
-        @Specialization
-        public Object sub(DynamicObject a, long b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).subtract(BigInteger.valueOf(b)));
-        }
-
-        @Specialization
-        public double sub(DynamicObject a, double b) {
-            return Layouts.BIGNUM.getValue(a).doubleValue() - b;
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public Object sub(DynamicObject a, DynamicObject b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).subtract(Layouts.BIGNUM.getValue(b)));
-        }
-
-    }
-
-    @CoreMethod(names = "*", required = 1)
-    public abstract static class MulNode extends BignumCoreMethodNode {
-
-        public static MulNode create() {
-            return MulNodeFactory.create(null);
-        }
-
-        public abstract Object executeMul(Object a, Object b);
-
-        @TruffleBoundary
-        @Specialization
-        public Object mul(DynamicObject a, long b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).multiply(BigInteger.valueOf(b)));
-        }
-
-        @Specialization
-        public double mul(DynamicObject a, double b) {
-            return Layouts.BIGNUM.getValue(a).doubleValue() * b;
-        }
-
-        @TruffleBoundary
-        @Specialization(guards = "isRubyBignum(b)")
-        public Object mul(DynamicObject a, DynamicObject b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).multiply(Layouts.BIGNUM.getValue(b)));
-        }
-
-        @Specialization(guards = { "!isInteger(b)", "!isLong(b)", "!isDouble(b)", "!isRubyBignum(b)" })
-        public Object mul(DynamicObject a, Object b,
-                @Cached("createOnSelf()") CallDispatchHeadNode redoCoerced) {
-            return redoCoerced.call(null, a, "redo_coerced", coreStrings().MULTIPLY.getSymbol(), b);
-        }
-
-    }
-
-    @CoreMethod(names = "/", required = 1)
-    public abstract static class DivNode extends BignumCoreMethodNode {
-
-        public abstract Object executeDiv(VirtualFrame frame, Object a, Object b);
-
-        @TruffleBoundary
-        @Specialization
-        public Object div(DynamicObject a, long b) {
-            if (b == 0) {
-                throw new RaiseException(coreExceptions().zeroDivisionError(this));
-            }
-            final BigInteger bBigInt = BigInteger.valueOf(b);
-            final BigInteger aBigInt = Layouts.BIGNUM.getValue(a);
-            final BigInteger result = aBigInt.divide(bBigInt);
-            if (result.signum() == -1 && !aBigInt.mod(bBigInt.abs()).equals(BigInteger.ZERO)) {
-                return fixnumOrBignum(result.subtract(BigInteger.ONE));
-            } else {
-                return fixnumOrBignum(result);
-            }
-        }
-
-        @TruffleBoundary
-        @Specialization
-        public double div(DynamicObject a, double b) {
-            return Layouts.BIGNUM.getValue(a).doubleValue() / b;
-        }
-
-        @TruffleBoundary
-        @Specialization(guards = "isRubyBignum(b)")
-        public Object div(DynamicObject a, DynamicObject b) {
-            final BigInteger aBigInt = Layouts.BIGNUM.getValue(a);
-            final BigInteger bBigInt = Layouts.BIGNUM.getValue(b);
-            final BigInteger result = aBigInt.divide(bBigInt);
-            if (result.signum() == -1 && !aBigInt.mod(bBigInt.abs()).equals(BigInteger.ZERO)) {
-                return fixnumOrBignum(result.subtract(BigInteger.ONE));
-            } else {
-                return fixnumOrBignum(result);
-            }
-        }
-
-    }
-
-    // Defined in Java as we need to statically call #/
-    @CoreMethod(names = "div", required = 1)
-    public abstract static class IDivNode extends BignumNodes.BignumCoreMethodNode {
-
-        @Child private DivNode divNode = DivNodeFactory.create(null);
-        @Child private FloatNodes.FloorNode floorNode = FloatNodesFactory.FloorNodeFactory.create(null);
-
-        @Specialization
-        public Object idiv(VirtualFrame frame, Object a, Object b,
-                @Cached("createBinaryProfile()") ConditionProfile zeroProfile) {
-            Object quotient = divNode.executeDiv(frame, a, b);
-            if (quotient instanceof Double) {
-                if (zeroProfile.profile((double) b == 0.0)) {
-                    throw new RaiseException(coreExceptions().zeroDivisionError(this));
-                }
-                return floorNode.executeFloor((double) quotient);
-            } else {
-                return quotient;
-            }
-        }
-
-    }
-
-    @CoreMethod(names = {"%", "modulo"}, required = 1)
-    public abstract static class ModNode extends BignumCoreMethodNode {
-
-        @TruffleBoundary
-        @Specialization
-        public Object mod(DynamicObject a, long b) {
-            if (b == 0) {
-                throw new ArithmeticException("divide by zero");
-            } else if (b < 0) {
-                final BigInteger bigint = BigInteger.valueOf(b);
-                final BigInteger mod = Layouts.BIGNUM.getValue(a).mod(bigint.negate());
-                return fixnumOrBignum(mod.add(bigint));
-            }
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).mod(BigInteger.valueOf(b)));
-        }
-
-        @TruffleBoundary
-        @Specialization(guards = "isRubyBignum(b)")
-        public Object mod(DynamicObject a, DynamicObject b) {
-            final BigInteger bigint = Layouts.BIGNUM.getValue(b);
-            final int compare = bigint.compareTo(BigInteger.ZERO);
-            if (compare == 0) {
-                throw new ArithmeticException("divide by zero");
-            } else if (compare < 0) {
-                final BigInteger mod = Layouts.BIGNUM.getValue(a).mod(bigint.negate());
-                return fixnumOrBignum(mod.add(bigint));
-            }
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).mod(Layouts.BIGNUM.getValue(b)));
-        }
-
-        @Specialization(guards = {"!isInteger(b)", "!isLong(b)", "!isRubyBignum(b)"})
-        public Object mod(DynamicObject a, Object b,
-                @Cached("createOnSelf()") CallDispatchHeadNode redoCoerced) {
-            return redoCoerced.call(null, a, "redo_coerced", coreStrings().MODULO.getSymbol(), b);
-        }
-
-    }
-
-    @CoreMethod(names = "<", required = 1)
-    public abstract static class LessNode extends CoreMethodArrayArgumentsNode {
-
-        @Specialization
-        public boolean less(DynamicObject a, long b) {
-            return Layouts.BIGNUM.getValue(a).compareTo(BigInteger.valueOf(b)) < 0;
-        }
-
-        @Specialization
-        public boolean less(DynamicObject a, double b) {
-            return Double.compare(Layouts.BIGNUM.getValue(a).doubleValue(), b) < 0;
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public boolean less(DynamicObject a, DynamicObject b) {
-            return Layouts.BIGNUM.getValue(a).compareTo(Layouts.BIGNUM.getValue(b)) < 0;
-        }
-
-        @Specialization(guards = {
-                "!isRubyBignum(b)",
-                "!isInteger(b)",
-                "!isLong(b)",
-                "!isDouble(b)"})
-        public Object lessCoerced(DynamicObject a, Object b,
-                @Cached("createOnSelf()") CallDispatchHeadNode redoCompare) {
-            return redoCompare.call(null, a, "redo_compare", coreStrings().LESS_THAN.getSymbol(), b);
-        }
-
-    }
-
-    @CoreMethod(names = "<=", required = 1)
-    public abstract static class LessEqualNode extends CoreMethodArrayArgumentsNode {
-
-        @Specialization
-        public boolean lessEqual(DynamicObject a, long b) {
-            return Layouts.BIGNUM.getValue(a).compareTo(BigInteger.valueOf(b)) <= 0;
-        }
-
-        @Specialization
-        public boolean lessEqual(DynamicObject a, double b) {
-            return Layouts.BIGNUM.getValue(a).compareTo(BigInteger.valueOf((long) b)) <= 0;
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public boolean lessEqual(DynamicObject a, DynamicObject b) {
-            return Layouts.BIGNUM.getValue(a).compareTo(Layouts.BIGNUM.getValue(b)) <= 0;
-        }
-
-        @Specialization(guards = {
-                "!isRubyBignum(b)",
-                "!isInteger(b)",
-                "!isLong(b)",
-                "!isDouble(b)"})
-        public Object lessEqualCoerced(DynamicObject a, Object b,
-                @Cached("createOnSelf()") CallDispatchHeadNode redoCompare) {
-            return redoCompare.call(null, a, "redo_compare", coreStrings().LESS_OR_EQUAL.getSymbol(), b);
-        }
-    }
-
-    @CoreMethod(names = { "==", "===" }, required = 1)
-    public abstract static class EqualNode extends CoreMethodArrayArgumentsNode {
-
-        @Child private BooleanCastNode booleanCastNode;
-        @Child private CallDispatchHeadNode reverseCallNode;
-
-        @Specialization
-        public boolean equal(DynamicObject a, int b) {
-            return false;
-        }
-
-        @Specialization
-        public boolean equal(DynamicObject a, long b) {
-            return false;
-        }
-
-        @Specialization
-        public boolean equal(DynamicObject a, double b) {
-            return Layouts.BIGNUM.getValue(a).doubleValue() == b;
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public boolean equal(DynamicObject a, DynamicObject b) {
-            return Layouts.BIGNUM.getValue(a).equals(Layouts.BIGNUM.getValue(b));
-        }
-
-        @Specialization(guards = "!isRubyBignum(b)")
-        public boolean equal(VirtualFrame frame, DynamicObject a, DynamicObject b) {
-            if (booleanCastNode == null) {
-                CompilerDirectives.transferToInterpreterAndInvalidate();
-                booleanCastNode = insert(BooleanCastNodeGen.create(null));
-            }
-
-            if (reverseCallNode == null) {
-                CompilerDirectives.transferToInterpreterAndInvalidate();
-                reverseCallNode = insert(CallDispatchHeadNode.create());
-            }
-
-            final Object reversedResult = reverseCallNode.call(frame, b, "==", a);
-
-            return booleanCastNode.executeToBoolean(reversedResult);
-        }
-    }
-
-    @CoreMethod(names = ">=", required = 1)
-    public abstract static class GreaterEqualNode extends CoreMethodArrayArgumentsNode {
-
-        @Specialization
-        public boolean greaterEqual(DynamicObject a, long b) {
-            return Layouts.BIGNUM.getValue(a).compareTo(BigInteger.valueOf(b)) >= 0;
-        }
-
-        @Specialization
-        public boolean greaterEqual(DynamicObject a, double b) {
-            return Double.compare(Layouts.BIGNUM.getValue(a).doubleValue(), b) >= 0;
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public boolean greaterEqual(DynamicObject a, DynamicObject b) {
-            return Layouts.BIGNUM.getValue(a).compareTo(Layouts.BIGNUM.getValue(b)) >= 0;
-        }
-
-        @Specialization(guards = {
-                "!isRubyBignum(b)",
-                "!isInteger(b)",
-                "!isLong(b)",
-                "!isDouble(b)"})
-        public Object greaterEqualCoerced(DynamicObject a, Object b,
-                @Cached("createOnSelf()") CallDispatchHeadNode redoCompare) {
-            return redoCompare.call(null, a, "redo_compare", coreStrings().GREATER_OR_EQUAL.getSymbol(), b);
-        }
-    }
-
-    @CoreMethod(names = ">", required = 1)
-    public abstract static class GreaterNode extends CoreMethodArrayArgumentsNode {
-
-        @Specialization
-        public boolean greater(DynamicObject a, long b) {
-            return Layouts.BIGNUM.getValue(a).compareTo(BigInteger.valueOf(b)) > 0;
-        }
-
-        @Specialization
-        public boolean greater(DynamicObject a, double b) {
-            return Double.compare(Layouts.BIGNUM.getValue(a).doubleValue(), b) > 0;
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public boolean greater(DynamicObject a, DynamicObject b) {
-            return Layouts.BIGNUM.getValue(a).compareTo(Layouts.BIGNUM.getValue(b)) > 0;
-        }
-
-        @Specialization(guards = {
-                "!isRubyBignum(b)",
-                "!isInteger(b)",
-                "!isLong(b)",
-                "!isDouble(b)"})
-        public Object greaterCoerced(DynamicObject a, Object b,
-                @Cached("createOnSelf()") CallDispatchHeadNode redoCompare) {
-            return redoCompare.call(null, a, "redo_compare", coreStrings().GREATER_THAN.getSymbol(), b);
-        }
-    }
-
-    @CoreMethod(names = "~")
-    public abstract static class ComplementNode extends BignumCoreMethodNode {
-
-        @Specialization
-        public Object complement(DynamicObject value) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(value).not());
-        }
-
-    }
-
-    @CoreMethod(names = "&", required = 1)
-    public abstract static class BitAndNode extends BignumCoreMethodNode {
-
-        @Specialization
-        public Object bitAnd(DynamicObject a, long b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).and(BigInteger.valueOf(b)));
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public Object bitAnd(DynamicObject a, DynamicObject b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).and(Layouts.BIGNUM.getValue(b)));
-        }
-
-        @Specialization(guards = { "!isInteger(b)", "!isLong(b)", "!isRubyBignum(b)" })
-        public Object bitAnd(DynamicObject a, Object b,
-                @Cached("createOnSelf()") CallDispatchHeadNode redoCoerced) {
-            return redoCoerced.call(null, a, "redo_bit_coerced", coreStrings().AMPERSAND.getSymbol(), b);
-        }
-    }
-
-    @CoreMethod(names = "|", required = 1)
-    public abstract static class BitOrNode extends BignumCoreMethodNode {
-
-        @Specialization
-        public Object bitOr(DynamicObject a, long b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).or(BigInteger.valueOf(b)));
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public Object bitOr(DynamicObject a, DynamicObject b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).or(Layouts.BIGNUM.getValue(b)));
-        }
-
-        @Specialization(guards = { "!isInteger(b)", "!isLong(b)", "!isRubyBignum(b)" })
-        public Object bitAnd(DynamicObject a, Object b,
-                @Cached("createOnSelf()") CallDispatchHeadNode redoCoerced) {
-            return redoCoerced.call(null, a, "redo_bit_coerced", coreStrings().PIPE.getSymbol(), b);
-        }
-    }
-
-    @CoreMethod(names = "^", required = 1)
-    public abstract static class BitXOrNode extends BignumCoreMethodNode {
-
-        @Specialization
-        public Object bitXOr(DynamicObject a, long b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).xor(BigInteger.valueOf(b)));
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public Object bitXOr(DynamicObject a, DynamicObject b) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(a).xor(Layouts.BIGNUM.getValue(b)));
-        }
-
-        @Specialization(guards = { "!isInteger(b)", "!isLong(b)", "!isRubyBignum(b)" })
-        public Object bitAnd(DynamicObject a, Object b,
-                @Cached("createOnSelf()") CallDispatchHeadNode redoCoerced) {
-            return redoCoerced.call(null, a, "redo_bit_coerced", coreStrings().CIRCUMFLEX.getSymbol(), b);
-        }
-
-    }
-
-    @CoreMethod(names = "<<", required = 1, lowerFixnum = 1)
-    public abstract static class LeftShiftNode extends BignumCoreMethodNode {
-
-        public abstract Object executeLeftShift(DynamicObject a, Object b);
-
-        @TruffleBoundary
-        @Specialization
-        public Object leftShift(DynamicObject a, int b,
-                                @Cached("createBinaryProfile()") ConditionProfile bPositive) {
-            if (bPositive.profile(b >= 0)) {
-                return fixnumOrBignum(Layouts.BIGNUM.getValue(a).shiftLeft(b));
-            } else {
-                return fixnumOrBignum(Layouts.BIGNUM.getValue(a).shiftRight(-b));
-            }
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public Object leftShift(DynamicObject a, DynamicObject b,
-                                @Cached("create()") ToIntNode toIntNode) {
-            final BigInteger bBigInt = Layouts.BIGNUM.getValue(b);
-            if (bBigInt.signum() == -1) {
-                return 0;
-            } else {
-                // MRI would raise a NoMemoryError; JRuby would raise a coercion error.
-                return executeLeftShift(a, toIntNode.doInt(b));
-            }
-        }
-
-        @Specialization(guards = {"!isRubyBignum(b)", "!isInteger(b)", "!isLong(b)"})
-        public Object leftShift(DynamicObject a, Object b,
-                                @Cached("create()") ToIntNode toIntNode) {
-            return executeLeftShift(a, toIntNode.doInt(b));
-        }
-
-    }
-
-    @CoreMethod(names = ">>", required = 1, lowerFixnum = 1)
-    public abstract static class RightShiftNode extends BignumCoreMethodNode {
-
-        public abstract Object executeRightShift(DynamicObject a, Object b);
-
-        @Specialization
-        public Object rightShift(DynamicObject a, int b,
-                @Cached("createBinaryProfile()") ConditionProfile bPositive) {
-            if (bPositive.profile(b >= 0)) {
-                return fixnumOrBignum(Layouts.BIGNUM.getValue(a).shiftRight(b));
-            } else {
-                return fixnumOrBignum(Layouts.BIGNUM.getValue(a).shiftLeft(-b));
-            }
-        }
-
-        @Specialization
-        public Object rightShift(DynamicObject a, long b) {
-            assert !CoreLibrary.fitsIntoInteger(b);
-            return 0;
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public int rightShift(DynamicObject a, DynamicObject b) {
-            return 0;
-        }
-
-        @Specialization(guards = {"!isRubyBignum(b)", "!isInteger(b)", "!isLong(b)"})
-        public Object rightShift(DynamicObject a, Object b,
-                @Cached("create()") ToIntNode toIntNode) {
-            return executeRightShift(a, toIntNode.doInt(b));
-        }
-
-
-    }
-
-    @CoreMethod(names = { "abs", "magnitude" })
-    public abstract static class AbsNode extends BignumCoreMethodNode {
-
-        @Specialization
-        public Object abs(DynamicObject value) {
-            return fixnumOrBignum(Layouts.BIGNUM.getValue(value).abs());
-        }
-
-    }
-
-    @CoreMethod(names = "bit_length")
-    public abstract static class BitLengthNode extends CoreMethodArrayArgumentsNode {
-
-        @Specialization
-        public int bitLength(DynamicObject value) {
-            return Layouts.BIGNUM.getValue(value).bitLength();
-        }
-
-    }
-
-    @CoreMethod(names = "divmod", required = 1)
-    public abstract static class DivModNode extends CoreMethodArrayArgumentsNode {
-
-        @Child private GeneralDivModNode divModNode = new GeneralDivModNode();
-
-        @Specialization
-        public DynamicObject divMod(DynamicObject a, long b) {
-            return divModNode.execute(Layouts.BIGNUM.getValue(a), b);
-        }
-
-        @Specialization
-        public DynamicObject divMod(DynamicObject a, double b) {
-            return divModNode.execute(Layouts.BIGNUM.getValue(a), b);
-        }
-
-        @Specialization(guards = "isRubyBignum(b)")
-        public DynamicObject divMod(DynamicObject a, DynamicObject b) {
-            return divModNode.execute(Layouts.BIGNUM.getValue(a), Layouts.BIGNUM.getValue(b));
-        }
-
-    }
-
-    @CoreMethod(names = "even?")
-    public abstract static class EvenNode extends BignumCoreMethodNode {
-
-        @TruffleBoundary
-        @Specialization
-        public boolean even(DynamicObject value) {
-            return !Layouts.BIGNUM.getValue(value).testBit(0);
-        }
-
-    }
-
-    @CoreMethod(names = "hash")
-    public abstract static class HashNode extends CoreMethodArrayArgumentsNode {
-
-        private static final int CLASS_SALT = 64260138; // random number, stops hashes for similar values but different classes being the same, static because we want deterministic hashes
-
-        @Specialization
-        public long hash(DynamicObject value) {
-            return getContext().getHashing(this).hash(CLASS_SALT, Layouts.BIGNUM.getValue(value).hashCode());
-        }
-
-    }
-
-    @CoreMethod(names = "odd?")
-    public abstract static class OddNode extends BignumCoreMethodNode {
-
-        @TruffleBoundary
-        @Specialization
-        public boolean odd(DynamicObject value) {
-            return Layouts.BIGNUM.getValue(value).testBit(0);
-        }
-
-    }
-
-    @CoreMethod(names = "size")
-    public abstract static class SizeNode extends CoreMethodArrayArgumentsNode {
-
-        @Specialization
-        public int size(DynamicObject value) {
-            return (Layouts.BIGNUM.getValue(value).bitLength() + 7) / 8;
-        }
-
-    }
-
-    @CoreMethod(names = "to_f")
-    public abstract static class ToFNode extends CoreMethodArrayArgumentsNode {
-
-        @Specialization
-        public double toF(DynamicObject value) {
-            return Layouts.BIGNUM.getValue(value).doubleValue();
-        }
-
-    }
-
-    @CoreMethod(names = { "to_s", "inspect" }, optional = 1, lowerFixnum = 1)
-    public abstract static class ToSNode extends CoreMethodArrayArgumentsNode {
-
-        @Child private StringNodes.MakeStringNode makeStringNode = StringNodes.MakeStringNode.create();
-
-        @TruffleBoundary
-        @Specialization
-        public DynamicObject toS(DynamicObject value, NotProvided base) {
-            return makeStringNode.executeMake(Layouts.BIGNUM.getValue(value).toString(), USASCIIEncoding.INSTANCE, CodeRange.CR_7BIT);
-        }
-
-        @TruffleBoundary
-        @Specialization
-        public DynamicObject toS(DynamicObject value, int base) {
-            if (base < 2 || base > 36) {
-                throw new RaiseException(coreExceptions().argumentErrorInvalidRadix(base, this));
-            }
-
-            return makeStringNode.executeMake(Layouts.BIGNUM.getValue(value).toString(base), USASCIIEncoding.INSTANCE, CodeRange.CR_7BIT);
         }
 
     }
